[IMP] Bye to crm_rma_prodlot_supplier module: crm_rma_prodlot_supplier was joined with crm_rma_prodlot_invoice [IMP] do_detailed_transfer method from crm_rma_prodlot_supplier module is refactored to best method. It is joined with do_detailed_transfer from crm_rma_prodlot_invoice module
