+++ crm_claim_rma/__openerp__.py
@@ -43,7 +43,6 @@
         'default_warehouse_from_sale_team',
         # from crm_claim_rma
         'crm_rma_prodlot_invoice',
-        'crm_rma_prodlot_supplier',
     ],
     'data': [
         'data/ir_sequence_type.xml',
