+++ crm_rma_prodlot_invoice/i18n/crm_rma_prodlot_invoice.pot
@@ -24,3 +24,23 @@ msgstr ""
 #: model:ir.model,name:crm_rma_prodlot_invoice.model_sale_order_line
 msgid "Sales Order Line"
 msgstr ""
+
+#. module: crm_rma_prodlot_invoice
+#: model:ir.model,name:crm_rma_prodlot_invoice.model_stock_production_lot
+msgid "Lot/Serial"
+msgstr ""
+
+#. module: crm_rma_prodlot_invoice
+#: model:ir.model,name:crm_rma_prodlot_invoice.model_stock_transfer_details
+msgid "Picking wizard"
+msgstr ""
+
+#. module: crm_rma_prodlot_invoice
+#: field:stock.production.lot,supplier_id:0
+msgid "Supplier"
+msgstr ""
+
+#. module: crm_rma_prodlot_invoice
+#: help:stock.production.lot,supplier_id:0
+msgid "Supplier of good in claim"
+msgstr ""
