+++ crm_rma_prodlot_invoice/models/__init__.py
@@ -19,8 +19,8 @@
 #
 ##############################################################################
 
-from . import sale_order
-from . import sale_order_line
 from . import stock_production_lot
 from . import stock_transfer_details
 from . import stock_picking
+from . import sale_order
+from . import sale_order_line
