+++ crm_rma_prodlot_invoice/views/stock_production_lot.xml
@@ -7,6 +7,8 @@
             <field name="inherit_id" ref="stock.view_production_lot_form"/>
             <field name="arch" type="xml">
                 <xpath expr="//field[@name='product_id']" position="after">
+                    <field name="supplier_id" readonly="1"/>
+                    <field name="supplier_invoice_line_id" readonly="1"/>
                     <field name="invoice_line_id"/>
                 </xpath>
             </field>
