+++ crm_rma_prodlot_invoice/views/stock_transfer_details.xml
@@ -1,17 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
-        <record id="view_production_lot_form_supplier" model="ir.ui.view">
-            <field name="name">lot.supplier</field>
-            <field name="model">stock.production.lot</field>
-            <field name="inherit_id" ref="stock.view_production_lot_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='product_id']" position="after">
-                    <field name="supplier_id" readonly="1"/>
-                    <field name="supplier_invoice_line_id" readonly="1"/>
-                </xpath>
-            </field>
-        </record>
         <record id="view_lot_form_supplier_context" model="ir.ui.view">
             <field name="name">lot.supplier.context</field>
             <field name="model">stock.transfer_details</field>
