--- crm_rma_prodlot_supplier/__openerp__.py
@@ -1,38 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright 2015 Vauxoo
-#    Author: Yanina Aular, Osval Reyes
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-{
-    'name': 'Claim Prodlot Supplier',
-    'version': '8.0.1.0.0',
-    'author': 'Vauxoo,Odoo Community Association (OCA)',
-    'website': 'http://www.vauxoo.com/',
-    'category': 'RMA',
-    'license': 'AGPL-3',
-    'depends': [
-        'stock',
-        'crm_claim',
-        'purchase',
-    ],
-    'data': [
-        'views/stock_view.xml',
-    ],
-    'installable': True,
-}
