--- crm_rma_prodlot_supplier/i18n/es.po
@@ -1,36 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* crm_rma_prodlot_supplier
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 8.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-07-17 19:57+0000\n"
-"PO-Revision-Date: 2015-07-17 19:57+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: crm_rma_prodlot_supplier
-#: model:ir.model,name:crm_rma_prodlot_supplier.model_stock_production_lot
-msgid "Lot/Serial"
-msgstr "Lote/No. de Serie"
-
-#. module: crm_rma_prodlot_supplier
-#: model:ir.model,name:crm_rma_prodlot_supplier.model_stock_transfer_details
-msgid "Picking wizard"
-msgstr "Asistente de albarán"
-
-#. module: crm_rma_prodlot_supplier
-#: field:stock.production.lot,supplier_id:0
-msgid "Supplier"
-msgstr "Proveedor"
-
-#. module: crm_rma_prodlot_supplier
-#: help:stock.production.lot,supplier_id:0
-msgid "Supplier of good in claim"
-msgstr "Proveedor del producto en la reclamación"
