--- crm_rma_prodlot_supplier/models/stock_production_lot.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright 2015 Vauxoo
-#    Author: Yanina Aular, Osval Reyes
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp import fields, models
-
-
-class StockProductionLot(models.Model):
-
-    _inherit = 'stock.production.lot'
-
-    supplier_id = fields.Many2one('res.partner', string='Supplier',
-                                  help="Supplier of good in claim")
-
-    supplier_invoice_line_id = \
-        fields.Many2one('account.invoice.line',
-                        string='Supplier Invoice Line',
-                        help="Supplier invoice with the "
-                             "purchase of goods sold to "
-                             "customer")
